SE_CRMSystem/requirements.txt
aiohttp==3.6.2
airtest==1.0.27
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.2.0
Automat==0.7.0
bcrypt==3.1.7
beautifulsoup4==4.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
comtypes==1.1.7
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
Django==1.11.23
django-multiselectfield==0.1.8
ExifRead==2.1.2
facebook-wda==0.3.9
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
hrpc==1.0.8
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
image==1.5.27
incremental==17.5.0
itchat==1.3.10
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10.1
kiwisolver==1.1.0
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mss==4.0.3
multidict==4.5.2
numpy==1.16.4
opencv-contrib-python==3.4.2.17
opencv-python==4.1.0.25
paramiko==2.6.0
parsel==1.5.2
pbr==5.4.3
pika==1.1.0
Pillow==6.1.0
pocoui==1.0.76
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyExecJS==1.5.1
PyHamcrest==1.9.0
pymongo==3.8.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.1.1
pypiwin32==223
pypng==0.0.20
PyQRCode==1.2.1
python-dateutil==2.8.0
pytz==2019.1
pywin32==224
pywinauto==0.6.3
queuelib==1.5.0
redis==3.3.11
removebg==0.3
requests==2.22.0
retry==0.9.2
Scrapy==1.7.3
scrapy-redis==0.6.8
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.6
sqlparse==0.3.0
stevedore==1.31.0
tornado==6.0.3
Twisted==19.7.0
typing-extensions==3.7.4
urllib3==1.25.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
w3lib==1.21.0
websocket-client==0.56.0
Werkzeug==0.16.0
wordcloud==1.5.0
yarl==1.3.0
you-get==0.4.1355
zope.interface==4.6.0
